Let `serialize_buffer` allocate memory
Passing in a range of memory when constructing a `serialize_buffer` with the
`take` action without a deleter expects the memory to be allocated with the
allocator.

All call sites where a `new T[]` has been passed with ownership transfer have
been converted to let serialize_buffer allocate the memory instead.