Fix build issues on MSVC and HPX master
The master branch of HPX appears to have deprecated `get_gid()` and has
hidden it behind a macro that is not defined for some configurations.
All uses of `get_gid()` for CUDA and core have been replaced with
`get_id()`.

FindNVRTC looks in insufficient numbers of directories for the Win64
library, this augments the search paths.

The examples pull in `<unistd.h>` without reason, remove it.